requirements.txt
pymupdf>=1.23.0
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
